+++ src/Local.h
@@ -16,20 +16,20 @@ private:
 	long long id;
 	pair<float, float> coordinates;		//latitude, longitude
 	float diffDistCenter;
-	int height = 0;
-	int diffHeightCenter = 0;
+	float height = 0;
+	float diffHeightCenter = 0;
 	map<long long, string> roads;		//idRoad, roadName
 	pair<bool, int> sharingPoint = pair<bool, int>(false, 0);	// is/is not sharingpoint, spots available
 	bool userLocation = false;
 	bool cityCenter = false;
 public:
 	Local(){};
-	Local(long long id, pair<float, float> coord, int height);
+	Local(long long id, pair<float, float> coord, float height);
 	long long getId() const;
 	pair<float, float> getCoordinates() const;
 	float getDiffDistCenter() const;
-	int getHeight() const;
-	int getDiffHeightCenter() const;
+	float getHeight() const;
+	float getDiffHeightCenter() const;
 	float getWeight(const Local nd, searchOptions op) const;
 	pair<bool, int> getSharingPoint();
 	bool getUserLocation();
@@ -40,7 +40,7 @@ public:
 	void setSharingPoint(pair<bool, int> sp);
 	void setUserLocation();
 	void setCityCenter();
-	void setDiffHeightCenter(int diff);
+	void setDiffHeightCenter(float diff);
 	void setDiffDistCenter(float diff);
 	void addRoad(pair<long long, string> road);
 	bool setRoadName(long long roadId, string roadName);
